feat(client): login google 외부 주입으로 변경
