+++ apps/client/src/app/auth/page.tsx
@@ -2,12 +2,28 @@
 
 import styled from '@emotion/styled';
 import { Spinner } from '@sickgyun/ui';
+import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
+import { LOCAL_STORAGE_KEY } from '@/constants/storage';
 import { useLoginGoogle } from '@/hooks/api/auth/useLoginGoogle';
+import { Storage } from '@/libs/api/storage';
 import { getGoogleAccessToken } from '@/utils/getGoogleAccessToken';
 
 const GoogleLoginPage = () => {
-  const { mutate: loginGoogleMutate } = useLoginGoogle();
+  const router = useRouter();
+  const { mutate: loginGoogleMutate } = useLoginGoogle({
+    onSuccess: (data) => {
+      const { accessToken, refreshToken } = data;
+
+      Storage.setItem(LOCAL_STORAGE_KEY.accessToken, `Bearer ${accessToken}`);
+      Storage.setItem(LOCAL_STORAGE_KEY.refreshToken, `Bearer ${refreshToken}`);
+      router.replace('/');
+    },
+    onError: () => {
+      alert('학교 계정으로 로그인 해주세요.');
+      router.replace('/');
+    },
+  });
 
   useEffect(() => {
     const accessToken = getGoogleAccessToken();
