+++ apps/client/src/hooks/api/auth/useLoginGoogle.ts
@@ -1,9 +1,7 @@
+import type { UseMutationOptions } from '@tanstack/react-query';
 import { useMutation } from '@tanstack/react-query';
 import type { AxiosError } from 'axios';
-import { useRouter } from 'next/navigation';
-import { LOCAL_STORAGE_KEY } from '@/constants/storage';
 import { post } from '@/libs/api/client';
-import { Storage } from '@/libs/api/storage';
 import type { ApiErrorScheme } from '@/libs/exceptions';
 
 type LoginGoogleRequest = {
@@ -17,25 +15,16 @@ type LoginGoogleResponse = {
 
 export const GOOGLE_LOGIN_MUTATION_KEY = 'googleLogin';
 
-export const useLoginGoogle = () => {
-  const router = useRouter();
-
-  return useMutation<LoginGoogleResponse, AxiosError<ApiErrorScheme>, LoginGoogleRequest>(
-    {
-      mutationKey: [GOOGLE_LOGIN_MUTATION_KEY],
-      mutationFn: (data) => post('/auth/login', data),
-      onSuccess: (data) => {
-        const { accessToken, refreshToken } = data;
-
-        Storage.setItem(LOCAL_STORAGE_KEY.accessToken, `Bearer ${accessToken}`);
-        Storage.setItem(LOCAL_STORAGE_KEY.refreshToken, `Bearer ${refreshToken}`);
-
-        router.replace('/');
-      },
-      onError: () => {
-        alert('학교 계정으로 로그인 해주세요.');
-        router.replace('/');
-      },
-    }
-  );
+export const useLoginGoogle = (
+  options: UseMutationOptions<
+    LoginGoogleResponse,
+    AxiosError<ApiErrorScheme>,
+    LoginGoogleRequest
+  >
+) => {
+  return useMutation({
+    mutationKey: [GOOGLE_LOGIN_MUTATION_KEY],
+    mutationFn: (data) => post('/auth/login', data),
+    ...options,
+  });
 };
